feat: move kubernetes-related functions to library
