+++ ftlib/utils/kubernetes/__init__.py
@@ -1,0 +1,1 @@
+from ftlib.utils.kubernetes.svc import get_peer_set  # noqa: F401
