+++ test/kubernetes/script/main.py
@@ -3,7 +3,6 @@ from __future__ import print_function
 import argparse
 import logging
 import os
-import socket
 import time
 
 import numpy as np
@@ -13,6 +12,7 @@ import torch.nn as nn
 import torch.optim as optim
 
 from ftlib import BasicFTLib
+from ftlib.utils.kubernetes import get_peer_set
 
 LOGLEVEL = os.environ.get("LOGLEVEL", "WARNING").upper()
 logging.basicConfig(level=LOGLEVEL)
@@ -36,13 +36,6 @@ parser.add_argument(
     default=1000,
     help="number of samples in dummy dataset",
 )
-
-
-def get_peer_set(svc_name):
-    my_ip = socket.gethostbyname(socket.gethostname())
-    temp_set = socket.getaddrinfo(svc_name, 0, proto=socket.IPPROTO_TCP)
-    peer_set = {peer[-1][0] for peer in temp_set if peer[-1][0] != my_ip}
-    return peer_set
 
 
 class SyntheticData(torch.utils.data.Dataset):
